requirements.txt
dnspython
elasticsearch
netifaces
pytz
requests
tzlocal
smbus2
